% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valmyndigheten_get.R
\name{valmyndigheten_get}
\alias{valmyndigheten_get}
\title{valmyndigheten_get}
\usage{
valmyndigheten_get(years, election_types)
}
\arguments{
\item{years}{A numeric vector of years (2010, 2014 and/or 2018)}

\item{election_types}{A character vector of election types ("riksdag", "county" and/or "municipality")}
}
\value{
A tibble of elections matching the specificed \code{year} and \code{election_type}
}
\description{
Get data from Valmyndigheten
}
\examples{
valmyndigheten_get(c(2010), c("riksdag"))
valmyndigheten_get(c(2010, 2014, 2018), c("riksdag", "county", "municipal"))
}
